@@ -1,6 +1,6 @@
 package com.group.libraryapp.temp;
 
-import javax.persistence.*;
+import jakarta.persistence.*;
 
 @Entity
 public class Address {
