@@ -1,32 +1,9 @@
 package com.group.libraryapp.domain.user;
 
-//public class User {
-//  private String name;
-//
-//  private Integer age;
-//
-//  public User(String name, Integer age) {
-//    if (name == null || name.isBlank()) {
-//      throw new IllegalArgumentException(String.format("잘못된 name(%s)이 들어왔습니다", name));
-//    }
-//    this.name = name;
-//    this.age = age;
-//  }
-//
-//  public String getName() {
-//    return name;
-//  }
-//
-//  public Integer getAge() {
-//    return age;
-//  }
-//
-//}
-//
-
 import com.group.libraryapp.domain.user.loanhistory.UserLoanHistory;
+import jakarta.persistence.*;
 
-import javax.persistence.*;
+//import javax.persistence.*;
 import java.util.ArrayList;
 import java.util.List;
 
@@ -35,8 +12,10 @@ public class User {
   @Id
   @GeneratedValue(strategy = GenerationType.IDENTITY)
   private Long id = null;
+
   @Column(nullable = false, length = 20)
   private String name;
+
   private Integer age;
 
   @OneToMany(mappedBy = "user", cascade = CascadeType.ALL, orphanRemoval = true)
